Move server.go to api package
